pyproject.toml
[project]
name = "PokeBot"
version = "0.1.0"
description = ""
authors = [

]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "discord-py (>=2.5.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "setuptools (>=75.8.2,<76.0.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.9"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 149

[tool.ruff.lint]
preview = true
select = ['F', 'E', 'W', 'PL', 'PT']

[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "str"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'python src/main.py'
pre_test = 'task lint'
test = 'pytest -s -x -vv'
post_test = 'coverage html'

[tool.pytest.ini_options]
pythonpath = [".", "src"]
addopts = '-p no:warnings'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
